aws/requirements.txt
aws-cdk-lib==2.35.0
aws-cdk.aws-lambda-python-alpha
constructs>=10.0.0,<11.0.0
